requirements.txt
asgiref==3.8.1
Django==5.2
pillow==11.1.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
requests==2.32.3
gunicorn==22.0.0
